Update annotations and version in conf files
Modified the wording for annotations in PostInject and PostConstruct files to match the actual behavior of the methods they annotate. Updated the version in build.gradle and cleaned up the format. Also, adjusted the settings and properties regarding the change tracking in the .idea files, and added credentials for Fumaz repository. These updates are necessary for correctly reflecting the status and functionality of the codebase and properly tracking the changes being made.